refactor: make language server code reusable by catalyst
